+++ frontend/src/Components/ExercisesCreation/exercisesCreation.css
@@ -1,0 +1,77 @@
+main.add-exercise-container {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    margin: 50px;
+}
+h2.add-title-exercise {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    margin: 10px;
+}
+
+.form-add-exercise {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+}
+
+.add-ex {
+    margin-bottom: 5px;
+    margin-top: 15px;
+    margin-right: 10px;
+    align-items: center;
+    display: flex;
+    justify-content: center;
+}
+
+.add-name-form,
+.add-description-form,
+.add-typology-form,
+.add-muscleGroup-form,
+.add-pic-form {
+    border: 1px solid purple;
+    border-radius: 4px;
+    width: 300px;
+    padding: 6px;
+    font-size: 14px;
+    margin-left: 5px;
+    align-items: center;
+}
+
+textarea {
+    font-size: 12px;
+    height: 150px;
+    width: 300px;
+}
+
+.ex-button-add {
+    border: 0 solid;
+    box-shadow: inset 0 0 20px rgba(255, 255, 255, 0);
+    outline: 1px solid;
+    padding: 0.1rem 2rem;
+    outline-color: rgba(255, 255, 255, 0.5);
+    outline-offset: 0px;
+    text-shadow: none;
+    transition: all 1250ms cubic-bezier(0.19, 1, 0.22, 1);
+    background-color: #7b468c;
+    font-size: 15px;
+    color: white;
+    margin-right: 10px;
+    margin: 15px;
+}
+
+.ex-button-add {
+    display: flex;
+    align-items: center;
+    flex-direction: row;
+}
+
+/* Media queries para diseño responsive */
+
+@media screen and (min-width: 768px) {
+}
+
+@media screen and (min-width: 375px) {
+}
