+++ frontend/src/Components/ExercisesEdit_Mari&Aitor/exercisesEdit.css
@@ -1,0 +1,98 @@
+main.edit-form-container {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    border-top: 1px solid black;
+    margin: 50px;
+}
+h2.ex-title-edit {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    margin: 10px;
+}
+.form-edition-exercise {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+}
+
+.ex-edit {
+    margin-bottom: 5px;
+    margin-top: 15px;
+    margin-right: 10px;
+    align-items: center;
+    display: flex;
+    justify-content: center;
+}
+
+.edit-name-form,
+.edit-description-form,
+.edit-typology-form,
+.edit-muscleGroup-form {
+    border-bottom: 1px solid purple;
+    border-radius: 4px;
+    width: 300px;
+    padding: 6px;
+    font-size: 14px;
+    margin-left: 5px;
+    align-items: center;
+}
+
+.edit-pic,
+.edit-pic-2 {
+    height: 200px;
+    width: 200px;
+    margin-left: 80px;
+}
+
+textarea {
+    font-size: 12px;
+    height: 150px;
+    width: 300px;
+}
+
+.edit-img {
+    margin-top: 20px;
+    border: 1px solid purple;
+    border-radius: 4px;
+    width: 100%;
+    padding: 6px;
+    font-size: 16px;
+    margin-bottom: 5px;
+}
+
+.edit-update,
+.edit-cancel {
+    border: 0 solid;
+    box-shadow: inset 0 0 20px rgba(255, 255, 255, 0);
+    outline: 1px solid;
+    padding: 0.1rem 2rem;
+    outline-color: rgba(255, 255, 255, 0.5);
+    outline-offset: 0px;
+    text-shadow: none;
+    transition: all 1250ms cubic-bezier(0.19, 1, 0.22, 1);
+    background-color: #7b468c;
+    font-size: 15px;
+    color: white;
+    margin-right: 10px;
+    margin: 15px;
+}
+
+.buttons-edit-form {
+    display: flex;
+    align-items: center;
+    flex-direction: row;
+}
+
+/* Media queries para diseño responsive */
+
+@media screen and (min-width: 768px) {
+}
+
+@media screen and (min-width: 375px) {
+    .edit-pic,
+    .edit-pic-2 {
+        margin-left: 60px;
+    }
+}
