+++ frontend/src/Components/Shared/Input/input.css
@@ -15,7 +15,7 @@
   color: gray;
   margin-right: 150px;
 } */
-input {
+/* input {
   font-size: 1px;
   padding: 5px 5px 5px 2.5px;
   display: inline-block;
@@ -45,4 +45,4 @@ label {
   position: relative;
   display: block;
   width: 250px;
-}
+} */
