+++ frontend/src/Pages/editProfile/editprofile.css
@@ -37,13 +37,16 @@ main.edit-profile {
     outline-offset: 5px;
     text-shadow: 1px 1px 2px #427388;
 }
-input {
+input #edit-avatar,
+#edit-name,
+#edit-surname,
+#edit-email {
     font-size: 12px;
     padding: 5px 5px 5px 2.5px;
     display: inline-block;
     width: 150px;
     border: none;
-    border-bottom: 1px solid #515151;
+    border-bottom: 1px solid purple;
     background: transparent;
     color: #333;
 }
